[MARLIN-244] Decorate SynchronizationController
